Updated check_operations
Cleaned out main a little.  Check operations is now called from analysis for one link at a time. It doesn't take head reference anymore because label error checking is now separate.  Match labels is the new function that uses previous check label to find out if the labels are correct. Now shows on which row the error occurred and unless it's a label, shows also what instruction was the issue, also takes into account if there are too many or too few instructions for the operation.